+++ app/models/volunteer.rb
@@ -13,7 +13,7 @@ class Volunteer < ActiveRecord::Base
   end
   
   def after_destroy
-    person.destroy
+    person.destroy if person
   end
   
   # Validates fields
